Extract getRandomString so it can be used in other tests.
